+++ jmh/src/test/java/com/uber/nullaway/jmh/BenchmarkCompilationTest.java
@@ -3,6 +3,7 @@ package com.uber.nullaway.jmh;
 import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
+import org.junit.Ignore;
 import org.junit.Test;
 
 /** Tests that all our JMH benchmarks compile successfully */
@@ -14,6 +15,7 @@ public class BenchmarkCompilationTest {
   }
 
   @Test
+  @Ignore("Need to update caffeine version; https://github.com/uber/NullAway/issues/1108")
   public void testCaffeine() throws IOException {
     assertTrue(new CaffeineCompiler().compile());
   }