+++ nullaway/src/test/java/com/uber/nullaway/AccessPathsTests.java
@@ -79,11 +79,17 @@ public class AccessPathsTests extends NullAwayTestsBase {
             "import javax.annotation.Nullable;",
             "public class Test {",
             "  private static final int INT_KEY = 42;", // Guaranteed constant!
+            "  // Guaranteed constant after class loading",
+            "  private static final int INT_KEY_HC = \"teststr\".hashCode();",
             "  public void testEnhancedFor(NullableContainer<String, NullableContainer<Integer, Object>> c) {",
             "    if (c.get(\"KEY_STR\") != null && c.get(\"KEY_STR\").get(INT_KEY) != null) {",
             "      c.get(\"KEY_STR\").get(INT_KEY).toString();",
             "      c.get(\"KEY_STR\").get(Test.INT_KEY).toString();",
             "      c.get(\"KEY_STR\").get(42).toString();", // Extra magic!
+            "    }",
+            "    if (c.get(\"KEY_STR\") != null && c.get(\"KEY_STR\").get(INT_KEY_HC) != null) {",
+            "      c.get(\"KEY_STR\").get(INT_KEY_HC).toString();",
+            "      c.get(\"KEY_STR\").get(Test.INT_KEY_HC).toString();",
             "    }",
             "  }",
             "}")
@@ -430,6 +436,60 @@ public class AccessPathsTests extends NullAwayTestsBase {
             "      // This is safe and handled",
             "      this.bar.toString();",
             "    }",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void mapKeysFromValueOf() {
+    defaultCompilationHelper
+        .addSourceLines(
+            "Foo.java",
+            "package com.uber;",
+            "import java.util.HashMap;",
+            "import java.util.Map;",
+            "public class Foo {",
+            "  private final Map<Integer, Object> map = new HashMap<>();",
+            "  private final Map<Long, Object> longMap = new HashMap<>();",
+            "  static Integer valueOf(int i) { return 0; }",
+            "  static Integer valueOf(int i, int j) { return i+j; }",
+            "  public void putThenGetIntegerValueOf() {",
+            "    map.put(Integer.valueOf(10), new Object());",
+            "    map.get(Integer.valueOf(10)).toString();",
+            "  }",
+            "  public void putThenGetLongValueOf() {",
+            "    longMap.put(Long.valueOf(10), new Object());",
+            "    longMap.get(Long.valueOf(10)).toString();",
+            "  }",
+            "  public void putThenGetFooValueOf() {",
+            "    map.put(valueOf(10), new Object());",
+            "    // Unknown valueOf method so we report a warning",
+            "    // BUG: Diagnostic contains: dereferenced expression map.get(valueOf(10)) is @Nullable",
+            "    map.get(valueOf(10)).toString();",
+            "    map.put(valueOf(10,20), new Object());",
+            "    // Unknown valueOf method so we report a warning",
+            "    // BUG: Diagnostic contains: dereferenced expression map.get(valueOf(10,20)) is @Nullable",
+            "    map.get(valueOf(10,20)).toString();",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void mapKeyFromIntegerValueOfStaticImport() {
+    defaultCompilationHelper
+        .addSourceLines(
+            "Foo.java",
+            "package com.uber;",
+            "import java.util.HashMap;",
+            "import java.util.Map;",
+            "import static java.lang.Integer.valueOf;",
+            "public class Foo {",
+            "  private final Map<Integer, Object> map = new HashMap<>();",
+            "  public void putThenGet() {",
+            "    map.put(valueOf(10), new Object());",
+            "    map.get(valueOf(10)).toString();",
             "  }",
             "}")
         .doTest();