+++ nullaway/src/test/java/com/uber/nullaway/jspecify/GenericsTests.java
@@ -2002,6 +2002,73 @@ public class GenericsTests extends NullAwayTestsBase {
         .doTest();
   }
 
+  @Test
+  public void issue1082() {
+    makeHelper()
+        .addSourceLines(
+            "Main.java",
+            "package com.uber;",
+            "import java.util.Optional;",
+            "public class Main {",
+            "  public interface Factory<T> {",
+            "    T create();",
+            "  }",
+            "  public interface Expiry<K, V> {}",
+            "  static class Config<K, V> {",
+            "    Config<K, V> setFactory(Optional<Factory<? extends Expiry<K, V>>> factory) {",
+            "      return this;",
+            "    }",
+            "  }",
+            "  static void caller(Config config) {",
+            "    // checking that we don't crash",
+            "    config.setFactory(Optional.<Object>empty());",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void issue1093() {
+    makeHelper()
+        .addSourceLines(
+            "Main.java",
+            "package com.uber;",
+            "import org.jspecify.annotations.Nullable;",
+            "public class Main {",
+            "    interface CacheLoader<V extends @Nullable Object> {",
+            "    }",
+            "    enum Loader implements CacheLoader<@Nullable Integer> {",
+            "        NULL;",
+            "        Loader() {}",
+            "    }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void nullUnmarkedGenericField() {
+    makeHelper()
+        .addSourceLines(
+            "Test.java",
+            "package com.unannotated;",
+            "import org.jspecify.annotations.*;",
+            "import java.util.function.Function;",
+            "public class Test {",
+            "  static Function<@Nullable String, @Nullable String> foo;",
+            "  @NullMarked",
+            "  static class Inner {",
+            "    static @Nullable Function<@Nullable String, @Nullable String> bar;",
+            "    void bar(Function<String,String> f) {",
+            "      // no error since foo is in @NullUnmarked code",
+            "      foo = f;",
+            "      // BUG: Diagnostic contains: Cannot assign from type Function<String, String>",
+            "      bar = f;",
+            "    }",
+            "  }",
+            "}")
+        .doTest();
+  }
+
   private CompilationTestHelper makeHelper() {
     return makeTestHelperWithArgs(
         Arrays.asList(